Refactor responder and rename to response factory
Refactor the responder and rename it to response factory since it's not itself a responder and the
existing pattern was just awful.
